MOPEDS

Documentation for MOPEDS.

All docstrings:

MOPEDS.BridgedWrapper

MOPEDS.Edge

MOPEDS.Node

MOPEDS.bridging_cost

MOPEDS.is_implemented

MOPEDS.required_funcs_with_argtypes

MOPEDS.static_which

MOPEDS.BridgedWrapper — Method

BridgedWrapper(obj)

A wrapper around obj with bridges all_bridges(obj) enabled. A BridgedWrapper can be used to efficiently query shortest paths.

MOPEDS.Edge — Type

Edge(bridge_index::Int, added_nodes::Vector{Node}, cost::Float64)

Hyper-edge in bridging hypergraph.

MOPEDS.Node — Type

Node(index::Int)

Simple node in bridging hypergraph.

MOPEDS.bridging_cost — Method

bridging_cost(
    bridge::AbstractBridge, func::Function, args_Type::Type{<:Tuple}
)

(Positive) cost of bridging with bridge.

MOPEDS.is_implemented — Method

is_implemented(
    bridge::AbstractBridge, func::Function, args_Type::Type{<:Tuple})

Trait function to indicate if a method exists.

MOPEDS.required_funcs_with_argtypes — Method

required_funcs_with_argtypes(
    bridge::AbstractBridge, func::Function, args_Type::Type{<:Tuple}
)

Return an indexable iterable of function-args_Type tuples required by bridge.

MOPEDS.static_which — Method

static_which(f, [type_tuple::Type{<:Tuple}, throw_error::Union{Val{true}, Val{false}}])

Similar to Base.which this returns the method of f that would be used by invoke for the given argument type tuple. If throw_error==Val(true), then an error is thrown if no suitable method is found. If throw_error==Val(false) and there is no method, then nothing is returned. Default is throw_error=Val(false).